requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
python-dotenv==0.19.0
bcrypt 